requirements.txt
pyfirmata
tk
matplotlib
numpy
argparse
scipy
pyserial
